Move exact_shap_ fixture into relevant sub-module.
